feature: rename prisma
